% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{penC}
\alias{penC}
\title{Solve the penalized logistic regression.}
\usage{
penC(x, y, off, beta, lam, pen)
}
\arguments{
\item{x}{samples of covariates which is a \eqn{n*p} matrix.}

\item{y}{samples of binary outcome which is a \eqn{n*1} vector.}

\item{off}{offset in logistic regression.}

\item{beta}{initial estimates.}

\item{lam}{value of the lasso penalty parameter \eqn{\lambda} for \eqn{\beta_1} and \eqn{\beta_2}.}

\item{pen}{1: MCP estimator; 2: SCAD estimator.}
}
\value{
A numeric vector, estimate of beta
}
\description{
Solve the penalized logistic regression.
}
